+++ .github/workflows/generate_artwork.R
@@ -32,6 +32,6 @@ artwork <- switch(type,
                   '18' = canvas_ribbons(colors = colorPalette('random', sample(3:6, size = 1)), background = colorPalette('random', 1)),
                   '19' = canvas_collatz(colors = colorPalette('random', n = 5), background = sample(c("black", "#fdf5e6", "#fafafa"), size = 1), n = sample(200:2000, size = 1), side = sample(c(TRUE, FALSE), size = 1)),
 				  '20' = canvas_watercolors(colors = colorPalette("random", n = sample(2:15, size = 1)), background = sample(c('#fafafa', "black", "#ebd5b3", "darkgoldenrod3", "lavenderblush2", "salmon1"), size = 1), layers = 50, depth = 3),
-				  '21' = canvas_flows(colors = colorPalette("random", n = sample(4:20, size = 1)), background = sample(c("#fafafa", "firebrick", "#f9f0e0", "black", "lavenderblush2"), size = 1), lines = sample(2000:3000, size = 1), iterations = 500, angles = sample(c("knn", "svm", "rf"), size = 1)))
+				  '21' = canvas_flow(colors = colorPalette("random", n = sample(4:20, size = 1)), background = sample(c("#fafafa", "firebrick", "#f9f0e0", "black", "lavenderblush2"), size = 1), lines = sample(2000:3000, size = 1), iterations = 500, angles.method = sample(c("knn", "svm", "rf"), size = 1)))
 
 saveCanvas(artwork, filename, width = ifelse(type == 19, yes = NA, no = 7), height = ifelse(type == 19, yes = NA, no = 7))