+++ R/saveCanvas.R
@@ -1,4 +1,4 @@
-#' Save a Canvas to an External Device.
+#' Save a Canvas to an External Device
 #'
 #' @description This function is a wrapper around \code{ggplot2::ggsave}. It provides a suggested export with square dimensions for a canvas created using the \code{aRtsy} package.
 #'